Todo list refactor
